Move all `blockUntil` to scalatestsuite
