+++ audio-api/src/test/scala/no/ndla/audioapi/service/search/AudioSearchServiceTest.scala
@@ -607,21 +607,4 @@ class AudioSearchServiceTest
     result2.results(4).title.language should be("nb")
     result2.results(5).title.language should be("pt-br")
   }
-
-  def blockUntil(predicate: () => Boolean): Unit = {
-    var backoff = 0
-    var done    = false
-
-    while (backoff <= 16 && !done) {
-      if (backoff > 0) Thread.sleep(200 * backoff)
-      backoff = backoff + 1
-      try {
-        done = predicate()
-      } catch {
-        case e: Throwable => println(("problem while testing predicate", e))
-      }
-    }
-
-    require(done, s"Failed waiting for predicate")
-  }
 }
