+++ audio-api/src/test/scala/no/ndla/audioapi/service/search/TagSearchServiceTest.scala
@@ -88,23 +88,6 @@ class TagSearchServiceTest extends IntegrationSuite(EnableElasticsearchContainer
     blockUntil(() => tagSearchService.countDocuments == tagsDistinctByLanguage.size)
   }
 
-  def blockUntil(predicate: () => Boolean): Unit = {
-    var backoff = 0
-    var done    = false
-
-    while (backoff <= 16 && !done) {
-      if (backoff > 0) Thread.sleep(200 * backoff)
-      backoff = backoff + 1
-      try {
-        done = predicate()
-      } catch {
-        case e: Throwable => println(("problem while testing predicate", e))
-      }
-    }
-
-    require(done, s"Failed waiting for predicate")
-  }
-
   test("That searching for tags returns sensible results") {
     val Success(result) = tagSearchService.matchingQuery("test", "nb", 1, 100)
