+++ concept-api/src/test/scala/no/ndla/conceptapi/service/search/DraftConceptSearchServiceTest.scala
@@ -780,21 +780,4 @@ class DraftConceptSearchServiceTest extends IntegrationSuite(EnableElasticsearch
     hits.head.copyright.head.origin should be(Some("Gotham City"))
     hits.head.copyright.head.creators.length should be(1)
   }
-
-  def blockUntil(predicate: () => Boolean): Unit = {
-    var backoff = 0
-    var done    = false
-
-    while (backoff <= 16 && !done) {
-      if (backoff > 0) Thread.sleep(200 * backoff)
-      backoff = backoff + 1
-      try {
-        done = predicate()
-      } catch {
-        case e: Throwable => println(("problem while testing predicate", e))
-      }
-    }
-
-    require(done, s"Failed waiting for predicate")
-  }
 }
