+++ draft-api/src/test/scala/no/ndla/draftapi/service/search/AgreementSearchServiceTest.scala
@@ -320,21 +320,4 @@ class AgreementSearchServiceTest extends IntegrationSuite(EnableElasticsearchCon
     scroll4.results.map(_.id) should be(expectedIds(4))
     scroll5.results.map(_.id) should be(List.empty)
   }
-
-  def blockUntil(predicate: () => Boolean): Unit = {
-    var backoff = 0
-    var done    = false
-
-    while (backoff <= 16 && !done) {
-      if (backoff > 0) Thread.sleep(200 * backoff)
-      backoff = backoff + 1
-      try {
-        done = predicate()
-      } catch {
-        case e: Throwable => println(("problem while testing predicate", e))
-      }
-    }
-
-    require(done, s"Failed waiting for predicate")
-  }
 }
