+++ draft-api/src/test/scala/no/ndla/draftapi/service/search/GrepCodesSearchServiceTest.scala
@@ -60,23 +60,6 @@ class GrepCodesSearchServiceTest extends IntegrationSuite(EnableElasticsearchCon
     blockUntil(() => grepCodesSearchService.countDocuments == allGrepCodesToIndex.size)
   }
 
-  def blockUntil(predicate: () => Boolean): Unit = {
-    var backoff = 0
-    var done    = false
-
-    while (backoff <= 16 && !done) {
-      if (backoff > 0) Thread.sleep(200 * backoff)
-      backoff = backoff + 1
-      try {
-        done = predicate()
-      } catch {
-        case e: Throwable => println(("problem while testing predicate", e))
-      }
-    }
-
-    require(done, s"Failed waiting for predicate")
-  }
-
   test("That searching for grepcodes returns sensible results") {
     val Success(result) = grepCodesSearchService.matchingQuery("KE", 1, 100)
