+++ draft-api/src/test/scala/no/ndla/draftapi/service/search/TagIndexServiceTest.scala
@@ -28,23 +28,6 @@ class TagIndexServiceTest extends IntegrationSuite(EnableElasticsearchContainer 
   override val converterService       = new ConverterService
   override val searchConverterService = new SearchConverterService
 
-  def blockUntil(predicate: () => Boolean): Unit = {
-    var backoff = 0
-    var done    = false
-
-    while (backoff <= 16 && !done) {
-      if (backoff > 0) Thread.sleep(200 * backoff)
-      backoff = backoff + 1
-      try {
-        done = predicate()
-      } catch {
-        case e: Throwable => println(("problem while testing predicate", e))
-      }
-    }
-
-    require(done, s"Failed waiting for predicate")
-  }
-
   test("That indexing does not fail if no tags are present") {
     tagIndexService.createIndexWithName(props.DraftTagSearchIndex)
