+++ search-api/src/test/scala/no/ndla/searchapi/TestData.scala
@@ -43,23 +43,6 @@ import java.time.LocalDateTime
 
 object TestData {
 
-  def blockUntil(predicate: () => Boolean): Unit = {
-    var backoff = 0
-    var done    = false
-
-    while (backoff <= 16 && !done) {
-      if (backoff > 0) Thread.sleep(200 * backoff)
-      backoff = backoff + 1
-      try {
-        done = predicate()
-      } catch {
-        case e: Throwable => println(("problem while testing predicate", e))
-      }
-    }
-
-    require(done, s"Failed waiting for predicate")
-  }
-
   private val publicDomainCopyright = Copyright("publicdomain", "", List(), List(), List(), None, None, None)
   private val byNcSaCopyright =
     Copyright("by-nc-sa", "Gotham City", List(Author("Writer", "DC Comics")), List(), List(), None, None, None)
